refactor: leveraged useReplayContext to get the setHover and setCurrentTime functions instead of prop drilling
