+++ static/app/views/replays/detail/console/index.tsx
@@ -4,7 +4,6 @@ import debounce from 'lodash/debounce';
 
 import CompactSelect from 'sentry/components/forms/compactSelect';
 import {Panel} from 'sentry/components/panels';
-import {relativeTimeInMs, showPlayerTime} from 'sentry/components/replays/utils';
 import SearchBar from 'sentry/components/searchBar';
 import {t} from 'sentry/locale';
 import space from 'sentry/styles/space';
@@ -23,12 +22,7 @@ interface Props {
 const getDistinctLogLevels = breadcrumbs =>
   Array.from(new Set<string>(breadcrumbs.map(breadcrumb => breadcrumb.level)));
 
-function Console({
-  breadcrumbs,
-  startTimestamp = 0,
-  setCurrentTime,
-  setCurrentHoverTime,
-}: Props) {
+function Console({breadcrumbs, startTimestamp = 0}: Props) {
   const [searchTerm, setSearchTerm] = useState('');
   const [logLevel, setLogLevel] = useState<BreadcrumbLevelType[]>([]);
   const handleSearch = debounce(query => setSearchTerm(query), 150);
@@ -67,19 +61,9 @@ function Console({
       {filteredBreadcrumbs.length > 0 ? (
         <ConsoleTable>
           {filteredBreadcrumbs.map((breadcrumb, i) => {
-            const diff = relativeTimeInMs(breadcrumb.timestamp || '', startTimestamp);
-            const onClick = () => setCurrentTime(diff);
-            const onMouseOver = () => setCurrentHoverTime(diff);
-            const onMouseOut = () => setCurrentHoverTime(undefined);
             return (
               <ConsoleMessage
-                onClick={onClick}
-                onMouseOver={onMouseOver}
-                onMouseOut={onMouseOut}
-                relativeTimestamp={showPlayerTime(
-                  breadcrumb.timestamp || '',
-                  startTimestamp
-                )}
+                startTimestamp={startTimestamp}
                 key={i}
                 isLast={i === breadcrumbs.length - 1}
                 breadcrumb={breadcrumb}
